fix - add struct.go contains all structures, lock before only set at cmd.go
